refactor(runtime): single-source the office-status transport contract (#122)
Eliminates the hand-maintained normalizePost.mjs runtime MIRROR (AVO-146 SITE 9). VALID_ROLES/STATUSES/MOODS, MAX_MOOD_DURATION, BLOCKED_REASONS, AGENT_CARRY_FIELDS, FIELD_SANITIZERS, sanitizeCarryFields and normalizePost now live ONCE in a new node-safe ESM module src/utils/statusContract.mjs (zero imports, .mjs so bare Node loads it directly). constants.js / classify.js / statusFields.js / normalizePost.js / normalizePost.mjs re-export from it — all existing import paths unchanged, server.mjs unchanged. classify.js keeps its rich BLOCKED_REASON_TABLE (UI metadata) plus a keys-equality guard against the contract codes.

AC-1 single source: adding a carry field is now one edit in statusContract.mjs. AC-2 bare Node: verified by booting server.mjs (POST returns {ok:true, agents:2}) and a new spawned-node import smoke in the drift-guard test. AC-3: the drift guard is reframed to a single-source wiring guard + the classify-table-keys guard; the AC-6 field-survival e2e is retained. Net -224 lines; behavior byte-identical on the A04 sanitization path (full suite 2029 tests green).
